+++ src/operationbot/eventListener.py
@@ -9,6 +9,7 @@ from discord.user import User
 
 from operationbot import config as cfg
 from operationbot import messageFunctions as msgFnc
+from operationbot import tasks
 from operationbot.bot import OperationBot
 from operationbot.errors import EventNotFound, RoleNotFound, RoleTaken, UnknownEmoji
 from operationbot.event import Event
@@ -41,6 +42,7 @@ class EventListener(Cog):
         print(msg)
         await commandchannel.send(msg)
         await self.bot.change_presence(activity=Game(name=cfg.GAME))
+        self.bot.loop.create_task(tasks.archive_past_events(self.bot))
         print("Logged in as", self.bot.user.name, self.bot.user.id)
         self.bot.processing = False